Flatten file hiearchy and use govendor deps
