+++ repo.go
@@ -1,12 +1,10 @@
-package user
-
-import "github.com/Sandy987/whatsforlunch/domain"
+package main
 
 // GetAllUsers returns all available User items
 func GetAllUsers() (Users, error) {
 	users := Users{}
 
-	err := domain.DB.Select(&users, "SELECT * from \"user\"")
+	err := DB.Select(&users, "SELECT * from \"user\"")
 	return users, err
 }
 
@@ -14,13 +12,13 @@ func GetAllUsers() (Users, error) {
 func GetUser(id int) (*User, error) {
 	user := User{}
 
-	err := domain.DB.Get(&user, "SELECT * from \"user\" WHERE ID = $1", id)
+	err := DB.Get(&user, "SELECT * from \"user\" WHERE ID = $1", id)
 	return &user, err
 }
 
 // CreateUser create a User object in the repository
 func CreateUser(u *HashedUser) {
-	domain.DB.NamedExec(`INSERT INTO \"user\" 
+	DB.NamedExec(`INSERT INTO \"user\" 
 		(
 			username,
 			first_name,
@@ -41,7 +39,7 @@ func CreateUser(u *HashedUser) {
 
 // UpdateUser updates an existing User
 func UpdateUser(u *User) {
-	domain.DB.NamedExec(`UPDATE \"user\" SET
+	DB.NamedExec(`UPDATE \"user\" SET
 		first_name = :first_name,
 		last_name = :last_name,
 	WHERE ID = :id`, u)
